refac(console): move task list view to `view` mod
